MoE common library
