+++ ttnn/cpp/ttnn/operations/ccl/all_to_all_combine/device/kernels/dataflow/writer_all_to_all_combine.cpp
@@ -3,42 +3,19 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "dataflow_api.h"
-#include "tt_metal/fabric/hw/inc/tt_fabric_api.h"
 #include "tt_metal/fabric/hw/inc/tt_fabric_interface.h"
+#include "tt_metal/fabric/hw/inc/tt_fabric_api.h"
 #include "tt_metal/fabric/hw/inc/edm_fabric/fabric_connection_manager.hpp"
+#include "ttnn/cpp/ttnn/operations/ccl/common/kernels/moe_utils.hpp"
 
 #include "../common.hpp"
-
-#include "dprint_pages.h"
 
 using tt::tt_fabric::NocUnicastAtomicIncCommandHeader;
 using tt::tt_fabric::NocUnicastCommandHeader;
 using tt::tt_fabric::WorkerToFabricEdmSender;
+using namespace ttnn::operations::ccl::common;
 
 namespace detail {
-
-// Batch is shared among all devices, or replicated along each row or each column.
-enum class ReplicateGroup : int {
-    NONE = -1,
-    ROWS = 1,
-    COLS = 0,
-};
-
-template <uint32_t src_chip_id, uint32_t mesh_cols, uint32_t mesh_rows, ReplicateGroup Axis>
-bool is_configured_target(uint32_t dest_chip_id) {
-    // axis is the direction along which we are allowed to send packets
-    // axis = 1; means we are allowed to send packets in the row direction
-    // axis = 0; means we are allowed to send packets in the column direction
-    // axis = -1; means we are allowed to send packets in all directions
-    if constexpr (Axis == ReplicateGroup::COLS) {  // check if they're on the same column
-        return src_chip_id % mesh_cols == dest_chip_id % mesh_cols;
-    } else if constexpr (Axis == ReplicateGroup::ROWS) {  // check if they're on the same row
-        return src_chip_id / mesh_cols == dest_chip_id / mesh_cols;
-    } else {
-        return true;  // if axis is not configured, we assume the target is configured, which is the default case, which
-                      // is all directions
-    }
-}
 
 template <uint32_t SourceChipId, uint32_t BatchSize, uint32_t MeshCols, uint32_t MeshRows, ReplicateGroup Axis>
 inline uint32_t get_device_idx_from_batch_idx(const uint32_t b) {
@@ -74,112 +51,9 @@ inline uint32_t get_output_page_idx(const uint32_t b, const uint32_t s, const ui
     const uint32_t bidx= b % batch_per_device;
     return k * batch_per_device *SeqSize + bidx*SeqSize+s;
 }
-
-// commonize me!
-template <size_t Size>
-inline void open_direction_connections(
-    const std::array<bool, Size>& directions,
-    std::array<WorkerToFabricEdmSender, Size>& connections,
-    size_t rt_args_idx) {
-    for (uint32_t i = 0; i < 4; i++) {
-        if (directions[i]) {
-            connections[i] =
-                tt::tt_fabric::WorkerToFabricEdmSender::build_from_args<ProgrammableCoreType::TENSIX>(rt_args_idx);
-            connections[i].open();
-        }
-    }
-}
-
-// commonize me!
-
-template <size_t Size>
-inline void close_direction_connections(
-    const std::array<bool, Size>& directions, std::array<WorkerToFabricEdmSender, Size>& connections) {
-    for (size_t i = 0; i < Size; ++i) {
-        if (directions[i]) {
-            connections[i].close();
-        }
-    }
-}
-
-// ! commonize me!
-/*
-enum eth_chan_directions {
-    EAST = 0,
-    WEST = 1,
-    NORTH = 2,
-    SOUTH = 3,
-    COUNT = 4,
-};*/
-
-template <uint32_t SrcChipId, uint32_t MeshCols, uint32_t MeshRows>
-inline eth_chan_directions get_direction(uint32_t dest_chip_id) {
-    // if along the same row, we go east or west
-    eth_chan_directions direction = eth_chan_directions::COUNT;
-
-    constexpr uint32_t Src_Row = SrcChipId / MeshCols, Src_Col = SrcChipId % MeshCols;
-    if (Src_Row == dest_chip_id / MeshCols) {
-        direction = SrcChipId < dest_chip_id ? eth_chan_directions::EAST : eth_chan_directions::WEST;
-    }
-    // if along the same column, we go north or south
-    else if (Src_Col == dest_chip_id % MeshCols) {
-        direction = SrcChipId > dest_chip_id ? eth_chan_directions::NORTH : eth_chan_directions::SOUTH;
-    }
-    // if not along the same row or column, we go north or south; north if dest_chip_id is smaller than src_chip_id
-    else {
-        direction = SrcChipId > dest_chip_id ? eth_chan_directions::NORTH : eth_chan_directions::SOUTH;
-    }
-    return direction;
-}
-
-template <uint32_t SrcChipId, uint32_t MeshCols, uint32_t MeshRows, int32_t MaxPacketSzBytes>
-inline void dispatch_input_remote_device(
-    const uint32_t dest_chip_id,
-    const uint32_t dest_mesh_id,
-    const uint32_t alignment,
-    int32_t size_bytes,
-    uint32_t input_token_read_addr,
-    uint64_t output_token_write_addr,
-    std::array<tt::tt_fabric::WorkerToFabricEdmSender, 4>& fabric_connections,
-    volatile PACKET_HEADER_TYPE* token_unicast_packet_header) {
-    // Clear the header buffer region.
-
-    const uint32_t route = get_direction<SrcChipId, MeshCols, MeshRows>(dest_chip_id);
-
-    // Populate packet header with routing information
-    zero_l1_buf((uint32_t*)token_unicast_packet_header, sizeof(PACKET_HEADER_TYPE));
-    fabric_set_unicast_route(
-        const_cast<LowLatencyMeshPacketHeader*>(token_unicast_packet_header),
-        static_cast<eth_chan_directions>(fabric_connections[route].direction),
-        SrcChipId,
-        dest_chip_id,
-        dest_mesh_id,
-        MeshCols);
-
-    while (size_bytes > 0) {
-        uint32_t curr_packet_size = std::min(MaxPacketSzBytes, size_bytes);
-
-        token_unicast_packet_header->to_noc_unicast_write(
-            NocUnicastCommandHeader{output_token_write_addr}, align(curr_packet_size, alignment));
-
-        fabric_connections[route].wait_for_empty_write_slot();
-
-        fabric_connections[route].send_payload_without_header_non_blocking_from_address(
-            input_token_read_addr, curr_packet_size);
-
-        fabric_connections[route].send_payload_flush_blocking_from_address(
-            (uint32_t)token_unicast_packet_header, sizeof(PACKET_HEADER_TYPE));
-
-        input_token_read_addr += curr_packet_size;
-        output_token_write_addr += curr_packet_size;
-        size_bytes -= curr_packet_size;
-    }
-}
 }  // namespace detail
 
 void kernel_main() {
-    using detail::ReplicateGroup;
-
     constexpr uint32_t metadata_cb_id = get_compile_time_arg_val(0);
     constexpr uint32_t local_experts_cb_id = get_compile_time_arg_val(1);
     constexpr uint32_t packet_header_cb_id = get_compile_time_arg_val(2);
@@ -216,7 +90,7 @@ void kernel_main() {
 
     const uint32_t rt_arg_count = 2;
     std::array<WorkerToFabricEdmSender, Num_Directions> fabric_connections;
-    detail::open_direction_connections(directions, fabric_connections, rt_arg_count);
+    open_direction_connections(directions, fabric_connections, rt_arg_count);
 
     InterleavedAddrGen<output_is_dram> output_addrgen{
         .bank_base_address = output_base_addr, .page_size = data_size_bytes};
@@ -260,18 +134,16 @@ void kernel_main() {
                     noc_async_write(src_data_l1_ptr,output_noc_addr,data_size_bytes);
                     noc_async_write_barrier();
                 } else {
-                    const auto route = detail::get_direction<src_chip_id, mesh_cols, mesh_rows>(dest_chip_id);
                     const auto& dest_mesh_id = dest_mesh_ids[dest_device_idx];
-                    detail::
-                        dispatch_input_remote_device<src_chip_id, mesh_cols, mesh_rows, fabric_max_packet_size_bytes>(
-                            dest_chip_id,
-                            dest_mesh_id,
-                            alignment,
-                            data_size_bytes,
-                            src_data_l1_ptr,
-                            output_noc_addr,
-                            fabric_connections,
-                            packet_headers[0]);
+                    dispatch_input_remote_device<src_chip_id, mesh_cols, mesh_rows, fabric_max_packet_size_bytes>(
+                        dest_chip_id,
+                        dest_mesh_id,
+                        alignment,
+                        data_size_bytes,
+                        src_data_l1_ptr,
+                        output_noc_addr,
+                        fabric_connections,
+                        packet_headers[0]);
                 }
                 cb_pop_front(data_cb_id,1);
             }
@@ -288,9 +160,9 @@ void kernel_main() {
         if (dest_chip_id == src_chip_id) {
             noc_semaphore_inc(get_noc_addr(global_semaphore_addr), 1);
             noc_async_atomic_barrier();
-        } else if (detail::is_configured_target<src_chip_id, mesh_cols, mesh_rows, replicate_axis>(dest_chip_id)) {
+        } else if (is_configured_target<src_chip_id, mesh_cols, mesh_rows, replicate_axis>(dest_chip_id)) {
             const auto & dest_mesh_id = dest_mesh_ids[device_idx];
-            const auto route = detail::get_direction<src_chip_id, mesh_cols, mesh_rows>(dest_chip_id);
+            const uint32_t route = get_next_hop_router_direction(dest_mesh_id, dest_chip_id);
 
             packet_headers[1]->to_noc_unicast_atomic_inc(
                 tt::tt_fabric::NocUnicastAtomicIncCommandHeader{global_noc_semaphore_addr, 1, 32, true});
@@ -309,7 +181,7 @@ void kernel_main() {
         }
     }
 
-    detail::close_direction_connections(directions, fabric_connections);
+    close_direction_connections(directions, fabric_connections);
 
     auto semaphore_ptr = reinterpret_cast<volatile tt_l1_ptr uint32_t*>(global_semaphore_addr);
     noc_semaphore_wait(semaphore_ptr, replicate_group_devices);
